chore: 패키지 레벨 재조정, tobe 패키지 삭제
