--- src/main/java/nextstep/ladder/domain/NotFoundUserException.java
@@ -1,8 +1,0 @@
-package nextstep.ladder.domain;
-
-public class NotFoundUserException extends RuntimeException {
-
-    public NotFoundUserException(String message) {
-        super(message);
-    }
-}
