+++ src/main/java/nextstep/ladder/domain/ladders/IllegalDirectionException.java
@@ -1,0 +1,8 @@
+package nextstep.ladder.domain.ladders;
+
+public class IllegalDirectionException extends RuntimeException {
+
+    public IllegalDirectionException(String message) {
+        super(message);
+    }
+}
