+++ src/main/java/nextstep/ladder/domain/ladders/LineCreateStrategy.java
@@ -1,0 +1,11 @@
+package nextstep.ladder.domain.ladders;
+
+import nextstep.ladder.common.Direction;
+
+public interface LineCreateStrategy {
+    public Direction first();
+
+    public Direction body(Direction exDirection);
+
+    public Direction last(Direction exDirection);
+}
