+++ src/main/java/nextstep/ladder/domain/result/NotFoundUserException.java
@@ -1,0 +1,8 @@
+package nextstep.ladder.domain.result;
+
+public class NotFoundUserException extends RuntimeException {
+
+    public NotFoundUserException(String message) {
+        super(message);
+    }
+}
