+++ src/main/java/nextstep/ladder/domain/result/Point.java
@@ -1,0 +1,19 @@
+package nextstep.ladder.domain.result;
+
+import nextstep.ladder.common.Direction;
+
+public class Point {
+    private int index;
+
+    public Point(int index) {
+        this.index = index;
+    }
+
+    public void move(Direction direction) {
+        index += direction.getValue();
+    }
+
+    public int getIndex() {
+        return index;
+    }
+}
