--- src/main/java/nextstep/ladder/domain/tobe/Direction.java
@@ -1,17 +1,0 @@
-package nextstep.ladder.domain.tobe;
-
-public enum Direction {
-    LEFT(-1),
-    CENTER(0),
-    RIGHT(1);
-
-    private final int value;
-
-    Direction(int value) {
-        this.value = value;
-    }
-
-    public int getValue() {
-        return value;
-    }
-}
