--- src/main/java/nextstep/ladder/domain/tobe/IllegalDirectionException.java
@@ -1,8 +1,0 @@
-package nextstep.ladder.domain.tobe;
-
-public class IllegalDirectionException extends RuntimeException {
-
-    public IllegalDirectionException(String message) {
-        super(message);
-    }
-}
