--- src/main/java/nextstep/ladder/domain/tobe/LineCreateStrategy.java
@@ -1,9 +1,0 @@
-package nextstep.ladder.domain.tobe;
-
-public interface LineCreateStrategy {
-    public Direction first();
-
-    public Direction body(Direction exDirection);
-
-    public Direction last(Direction exDirection);
-}
