--- src/main/java/nextstep/ladder/domain/tobe/Point.java
@@ -1,17 +1,0 @@
-package nextstep.ladder.domain.tobe;
-
-public class Point {
-    private int index;
-
-    public Point(int index) {
-        this.index = index;
-    }
-
-    public void move(Direction direction) {
-        index += direction.getValue();
-    }
-
-    public int getIndex() {
-        return index;
-    }
-}
