+++ src/main/java/nextstep/ladder/domain/users/User.java
@@ -1,0 +1,21 @@
+package nextstep.ladder.domain.users;
+
+public class User {
+    private static final int MAX_LENGTH = 5;
+    private final String name;
+
+    public User(String name) {
+        validateName(name);
+        this.name = name;
+    }
+
+    private void validateName(String name) {
+        if (name.length() > MAX_LENGTH) {
+            throw new IllegalArgumentException("name length is more than " + MAX_LENGTH);
+        }
+    }
+
+    public String getName() {
+        return name;
+    }
+}
