+++ src/main/java/nextstep/ladder/domain/users/Users.java
@@ -1,0 +1,24 @@
+package nextstep.ladder.domain.users;
+
+import java.util.List;
+import java.util.stream.Stream;
+
+public class Users {
+    private List<User> users;
+
+    public Users(List<User> users) {
+        this.users = users;
+    }
+
+    public int size() {
+        return users.size();
+    }
+
+    public Stream<User> stream() {
+        return users.stream();
+    }
+
+    public User getUser(int index) {
+        return users.get(index);
+    }
+}
