+++ src/test/java/nextstep/ladder/domain/fixture/PrizesFixture.java
@@ -1,0 +1,15 @@
+package nextstep.ladder.domain.fixture;
+
+import nextstep.ladder.domain.prizes.Prize;
+import nextstep.ladder.domain.prizes.Prizes;
+
+import java.util.Arrays;
+
+public class PrizesFixture {
+    public static Prizes TestPrizes = new Prizes(
+            Arrays.asList(
+                    new Prize("꽝"),
+                    new Prize("5000"),
+                    new Prize("꽝"),
+                    new Prize("3000")));
+}
