+++ src/test/java/nextstep/ladder/domain/fixture/SeriesConnectLineCreateStrategy.java
@@ -1,0 +1,22 @@
+package nextstep.ladder.domain.fixture;
+
+import nextstep.ladder.common.Direction;
+import nextstep.ladder.domain.ladders.LineCreateStrategy;
+
+public class SeriesConnectLineCreateStrategy implements LineCreateStrategy {
+
+    @Override
+    public Direction first() {
+        return Direction.CENTER;
+    }
+
+    @Override
+    public Direction body(Direction exDirection) {
+        return Direction.LEFT;
+    }
+
+    @Override
+    public Direction last(Direction exDirection) {
+        return Direction.CENTER;
+    }
+}
